feat: adjust the code structure in extensions/github1s
