+++ extensions/github1s/src/helpers/func.ts
