+++ extensions/github1s/src/providers/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @file export providers
+ * @author fezhang
+ */
+
+export { GitHub1sFileSystemProvider } from './fileSystemProvider';
+export { GitHub1sFileSearchProvider } from './fileSearchProvider';
+export { GitHub1sTextSearchProvider } from './textSearchProvider';
+export { GitHub1sSubmoduleDecorationProvider } from './submoduleDecorationProvider';
